requirements.txt
Django
djangorestframework
markdown
django-filter
psycopg2
